% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_resid_versus_fit.R
\name{diag_resid_versus_fit}
\alias{diag_resid_versus_fit}
\title{Fitted values against residuals}
\usage{
diag_resid_versus_fit(fit)
}
\arguments{
\item{fit}{a model object - currently working for glm}
}
\description{
Plots residuals versus fitted values of a generalized linear model.
}
