% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envel_glm.R
\name{envel}
\alias{envel}
\title{Envelope plots of generalized linear model}
\usage{
envel(fit, B = 100)
}
\arguments{
\item{fit}{a model object - currently working for glm}

\item{B}{number of bootstrap iterations}
}
\description{
Draws the envelope plot for the residual distributions of a generalized linear model.
}
