% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envel_glm.R
\name{envel_glm}
\alias{envel_glm}
\title{Envelope plots of generalized linear model}
\usage{
envel_glm(fit, B = 100, resid_type = "response")
}
\arguments{
\item{fit}{a model object - currently working for glm}

\item{B}{number of bootstrap iterations}

\item{resid_type}{type of the residual to be used}
}
\description{
Draws the envelope plot for the residual distributions of a generalized linear model.
}
