% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-diag-glm.R
\name{plot.diag_glm}
\alias{plot.diag_glm}
\title{Plot diag}
\usage{
\method{plot}{diag_glm}(x, graphic = "residuals", ...)
}
\arguments{
\item{x}{object}

\item{graphic}{-}

\item{...}{other}
}
\description{
Plot diag
}
\seealso{
Other glm diagnostic functions: \code{\link{diag_data}}
}
