% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-diag-lm-glm.R
\name{plot.diag_lm}
\alias{plot.diag_lm}
\title{Create a specific diagnostic graphic for lm fits}
\usage{
\method{plot}{diag_lm}(x, graphic = "residuals", ...)
}
\arguments{
\item{x}{a tibble of class \code{diag_lm}.}

\item{graphic}{a string specifying the graphic to be plotted. The available
choices are 'residuals' (default), 'leverage', 'cook' and 'link'.}

\item{...}{futher arguments.}
}
\description{
\code{ggplot} is used to plot a specific diagnostic graphic. The choices are:
residuals plot, leverage mesuare plot, Cook's distance plot and z variable
plot (link function plot). See \emph{Details} for more information.
}
\details{
The graph created will depend on the class of the \code{model} object and
choice of the \code{graphic} argument. For \code{diag_lm} class, the option
\code{"residuals"} creates a residuals vs fitted values plot, the option
\code{"cook"} creates a Cook's distance vs index plot, and the option
\code{"leverage"} plots a leverage vs index plot.

For more information about the measures used to create the plots, see the
\code{\link{diag_data}} function documentation.
}
\seealso{
Other glm diagnostic functions: \code{\link{diag_data}},
  \code{\link{plot.diag_glm}}
}
