% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit.R
\name{refit}
\alias{refit}
\title{Refit a model}
\usage{
refit(model, y)
}
\arguments{
\item{model}{a model object - currently working for lm and glm}

\item{y}{new response variable}
}
\description{
Refits a model changing the response variable
}
