% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag-data.R
\name{diag_data}
\alias{diag_data}
\title{Create a tibble with useful diagnostic measures}
\usage{
diag_data(model)
}
\arguments{
\item{model}{a previously fitted model of class \code{lm} or \code{glm}}
}
\value{
A tibble of class \code{diag_lm} or \code{diag_glm} (depending on the
  class of the \code{model} object) with the following columns:

  \itemize{ \item \strong{.rownames}: the row labels taken from the dataset
  \item the variables (response and predictors) used to fit the model \item
  \strong{.fitted}: predicted value \item \strong{.se.fit}: standard-error
  \item \strong{.resid}: residuals \item \strong{.hat}: h-measure \item
  \strong{.sigma}: \item \strong{.cooksd}: Cook's distance \item
  \strong{.std.resid}: standardized residuals \item \strong{.w}: weights
  \item \strong{.eta}: \item \strong{.pearson}: person residuals \item
  \strong{.z}: z-measure }
}
\description{
Uses broom::augment and some other useful measures
}
\examples{
fit_lm <- lm(mpg ~ cyl, data = mtcars)
diag_data(fit_lm)

fit_glm <- glm(mpg ~ cyl, family = Gamma(link = "log"), data = mtcars)
diag_data(fit_glm)

}
\seealso{
Other glm diagnostic functions: \code{\link{plot.diag_glm}},
  \code{\link{plot.diag_lm}}
}
