% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag-data.R
\name{diag_data}
\alias{diag_data}
\title{Create tibble with useful diagnostic measures}
\usage{
diag_data(model)
}
\arguments{
\item{model}{}
}
\description{
Uses broom::augment and some other useful measures
}
