% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_resid_versus_fit.R
\name{plot.diag_glm}
\alias{plot.diag_glm}
\title{Plot diag}
\usage{
\method{plot}{diag_glm}(x, graphic = "residuals", ...)
}
\arguments{
\item{x}{object}

\item{graphic}{-}

\item{...}{other}
}
\description{
Plot diag
}
